test(gleif): split client_test.go under 500 LOC (deslop #39)
Pure move of the four client-level cache tests (TestCacheHitMiss,
TestCacheDisabled, TestCacheStats, TestCacheClear) into a new
client_cache_test.go. No body changes.

client_test.go: 657 -> 481 LOC. client_cache_test.go: 186 LOC.
isAPIError helper stays in client_test.go (used by the validation/client
tests that remain there); the moved cache tests don't reference it.
